+++ samples/Issues/ShowRepoIssues.hs
@@ -1,21 +1,42 @@
-module ShowRepoIssue where
+{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE OverloadedStrings #-}
 
-import qualified Github.Issues as Github
-import Data.List (intercalate)
+import Data.Foldable (toList)
+import Data.List     (intercalate)
+import Data.Vector   (Vector)
 
+import qualified GitHub as Github
+
+main :: IO ()
 main = do
-  let limitations = [Github.OnlyClosed, Github.Mentions "mike-burns", Github.AssignedTo "jyurek"]
-  possibleIssues <- Github.issuesForRepo "thoughtbot" "paperclip" limitations
-  case possibleIssues of
-       (Left error) -> putStrLn $ "Error: " ++ show error
-       (Right issues) ->
-         putStrLn $ intercalate "\n\n" $ map formatIssue issues
+  let filt = Github.stateClosed <> Github.optionsMentioned "mike-burns" <> Github.optionsAssignee "jyurek"
+  printIssues =<< do
+    Github.github' $ Github.issuesForRepoR "thoughtbot" "paperclip" filt Github.FetchAll
 
-formatIssue issue =
-  (Github.githubOwnerLogin $ Github.issueUser issue) ++
-    " opened this issue " ++
-    (show $ Github.fromDate $ Github.issueCreatedAt issue) ++ "\n" ++
-    (Github.issueState issue) ++ " with " ++
-    (show $ Github.issueComments issue) ++ " comments" ++ "\n\n" ++
-    (Github.issueTitle issue)
+  printIssues =<< do
+    Github.github' $ Github.issuesForRepoR "haskell-github" "playground" Github.stateClosed Github.FetchAll
 
+printIssues :: Either Github.Error (Vector Github.Issue) -> IO ()
+printIssues = \case
+  Left err ->
+    putStrLn $ "Error: " ++ show err
+  Right issues ->
+    putStrLn $ intercalate "\n\n" $ map formatIssue $ toList issues
+
+formatIssue :: Github.Issue -> String
+formatIssue issue = concat
+
+  [ show $ Github.simpleUserLogin $ Github.issueUser issue
+  , " opened this issue "
+  , show $ Github.issueCreatedAt issue
+  , ".\n"
+
+  , "It is currently "
+  , show $ Github.issueState issue
+  , maybe "" (\ r -> " with reason " ++ show r) $ Github.issueStateReason issue
+  , " with "
+  , show $ Github.issueComments issue
+  , " comments.\n\n"
+
+  , show $ Github.issueTitle issue
+  ]